backend/requirements.txt
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
strands-agents>=0.1.0
resend>=2.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
strands-agents-tools>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
supabase>=2.4.0
boto3>=1.34.0
python-dotenv>=1.0.0
pypdf>=4.2.0
duckduckgo-search>=6.2.0

# dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.4.0
mypy>=1.10.0
